+++ src/main/kotlin/io/verik/intellij/inspection/common/VerikInspectionProvider.kt
@@ -18,9 +18,10 @@ package io.verik.intellij.inspection.common
 
 import com.intellij.codeInspection.InspectionToolProvider
 import com.intellij.codeInspection.LocalInspectionTool
+import io.verik.intellij.inspection.inspection.ComAssignmentNotVarInspection
 import io.verik.intellij.inspection.inspection.FileAnnotationInspection
-import io.verik.intellij.inspection.inspection.ModuleNotClass
-import io.verik.intellij.inspection.inspection.ModuleNotObject
+import io.verik.intellij.inspection.inspection.ModuleNotClassInspection
+import io.verik.intellij.inspection.inspection.ModuleNotObjectInspection
 import io.verik.intellij.inspection.inspection.TopNotModuleInspection
 import io.verik.intellij.inspection.inspection.UnsupportedElementInspection
 import io.verik.intellij.inspection.inspection.UnsupportedModifierInspection
@@ -33,8 +34,9 @@ class VerikInspectionProvider : InspectionToolProvider {
             UnsupportedElementInspection::class.java,
             UnsupportedModifierInspection::class.java,
             TopNotModuleInspection::class.java,
-            ModuleNotClass::class.java,
-            ModuleNotObject::class.java
+            ModuleNotClassInspection::class.java,
+            ModuleNotObjectInspection::class.java,
+            ComAssignmentNotVarInspection::class.java
         )
     }
 }